move data and types to separate files
